pyproject.toml
[tool.poetry]
name = "pw-homework-module-13"
version = "0.1.0"
description = ""
authors = ["RuslanaPashkutska <ryslana.pashkytska@gmail.com>"]
packages = [{ include = "src" }]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
alembic = ">=1.16.2,<2.0.0"
dotenv = ">=0.9.9,<0.10.0"
pydantic-settings = ">=2.10.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
fastapi = ">=0.110.0,<1.0.0"
uvicorn = { version = ">=0.29.0,<1.0.0", extras = ["standard"] }
python-jose = { version = ">=3.3.0,<4.0.0", extras = ["cryptography"] }
passlib = { version = ">=1.7.4,<2.0.0", extras = ["bcrypt"] }
cloudinary = ">=1.37.0,<2.0.0"
aioredis = ">=2.0.1,<3.0.0"
redis = "^6.2.0"
fastapi-mail = "^1.5.0"
fastapi-limiter = "^0.1.6"
django-environ = "^0.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"